Miscellaneous updates, comments, linter fixes
