+++ main.go
@@ -13,7 +13,7 @@ func main() {
 	defer func() {
 		if err := recover(); err != nil {
 			red := color.New(color.FgHiRed)
-			red.Fprintf(os.Stderr, "Runtime error: %v\n", err)
+			_, _ = red.Fprintf(os.Stderr, "Runtime error: %v\n", err)
 			os.Exit(1)
 		}
 	}()
