+++ version/version.go
@@ -1,5 +1,6 @@
 package version
 
+//nolint:gochecknoglobals // These variables receive injected values at compile-time.
 var (
 	Version   = "undefined"
 	GitCommit = "undefined"
